requirements.txt
flask==3.0.0
python-dotenv==1.0.1
pymongo==4.10.1
dnspython==2.6.1
PyJWT==2.9.0
bcrypt==4.1.2
cloudinary==1.41.0
daytime==0.4.0
gunicorn==21.2.0
sendgrid==6.11.0
